Changed isPrime method to work- part 2

+++ app/src/main/java/edu/up/cs371/primetester/PrimeTester.java
@@ -21,7 +21,7 @@ public class PrimeTester {
         }
         else
         {
-            for(int i = 3; i*i <= n; i += 2)
+            for(long i = 3; i*i <= n; i += 2)
             {
                 {
                     if(n%i == 0)
